requirements.txt
flask
flask-cors
python-dotenv
pymongo
azure-storage-blob
